wip: cleanup and consolidation, renaming
