+++ src/components/inputs/Combobox/Combobox.scss
@@ -29,3 +29,225 @@
 		}
 	}
 }
+@import '../../../styles/_partials/index';
+
+.Combobox {
+	--inputHeight: 37px;
+	--optionHeight: 37px;
+	display: inline-flex;
+	width: 100%;
+	position: relative;
+	max-width: 100%;
+
+	&.Combobox__InputHeightLarge {
+		--inputHeight: 57px;
+	}
+
+	&.Combobox__OptionHeightLarge {
+		--optionHeight: 57px;
+	}
+
+	.Combobox__SearchIcon {
+		pointer-events: none;
+		position: absolute;
+		margin-left: 15px;
+		opacity: 0;
+		transition: opacity 200ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+	}
+
+	.Combobox__BasicInput {
+		margin-bottom: 0;
+		width: 100%;
+
+		input {
+			padding: 0;
+			height: var(--inputHeight);
+			transition: padding-left 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+			padding-left: 20px;
+			padding-right: 42px;
+		}
+	}
+
+	&.Combobox__Disabled {
+		filter: grayscale(100%);
+		opacity: 0.5;
+
+		* {
+			pointer-events: none;
+		}
+	}
+
+	.Combobox__CaretIcon, .Combobox__CloseIcon {
+		position: absolute;
+		right: 20px;
+	}
+
+	&:not(.Combobox__Open) {
+
+		&:focus,
+		&:focus-visible {
+			outline: none;
+		}
+	}
+
+	&.Combobox__Open {
+		z-index: 5;
+
+		.Combobox__SearchIcon {
+			opacity: 1;
+		}
+
+		.Combobox__BasicInput {
+			input {
+				padding-left: 45px;
+				border-radius: 4px 4px 0 0;
+			}
+		}
+
+		.Combobox_Options {
+			display: block;
+		}
+	}
+
+	.DownloadSmall {
+		width: 12px;
+		height: 12px;
+		align-self: center;
+		margin: 0 7px 0 0;
+
+		path {
+			@include __theme-fill($green-dark, $green);
+		}
+	}
+
+	.Combobox_OptionsContainer {
+		overflow: auto;
+		max-height: inherit;
+	}
+
+	.Combobox_Options {
+		position: absolute;
+		top: calc(var(--inputHeight) + 1px);
+		left: 0;
+		display: none;
+		width: 100%;
+		@include theme-input-background-color;
+		@include theme-input-border-box-shadow;
+		border-radius: 0 0 4px 4px;
+	}
+
+	.Combobox_OptionGroup {
+		height: 37px;
+		line-height: 37px;
+		text-align: left;
+		@include theme-color-gray-else-gray25;
+		@include theme-background-gray15-else-graydark50;
+		@include tracking(5);
+		font-weight: 700;
+		padding: 0 20px;
+		display: flex;
+
+		a {
+			margin-left: auto;
+			@include theme-color-gray-else-gray25;
+			display: flex;
+			align-items: center;
+
+			svg {
+				width: 6px;
+				height: 14px;
+				margin: -1px 0 0 10px;
+
+				path {
+					fill: $gray75;
+				}
+			}
+
+			&:hover {
+				@include theme-color-gray-else-white;
+
+				svg path {
+					@include theme-fill-gray-else-white;
+				}
+			}
+		}
+	}
+
+	.Combobox_Option {
+		display: flex;
+		align-items: center;
+		line-height: var(--optionHeight);
+		height: var(--optionHeight);
+		padding: 0 10px;
+
+		&:last-of-type {
+			border-radius: 0 0 4px 4px;
+		}
+
+		@include cursorPointer;
+
+		.Combobox__Right {
+			margin: 0 0 0 auto;
+
+			.Combobox__Check {
+				margin-left: 10px;
+			}
+		}
+
+		.Combobox__SecondaryText {
+			margin-left: auto;
+			color: $gray75;
+			font-weight: 300;
+		}
+
+		svg.Combobox__ItemIcon,
+		img {
+			margin: 0 10px 0 0;
+			width: auto;
+			flex-shrink: 0;
+		}
+
+		img {
+			height: 22px;
+			border-radius: 1px;
+
+			@include if-theme-dark() {
+				filter: brightness(0.44);
+			}
+		}
+
+		svg.Combobox__ItemIcon {
+			height: 27px;
+		}
+
+		.Combobox_Option_Label {
+			max-width: 100%;
+			overflow: hidden;
+			text-overflow: ellipsis;
+			white-space: nowrap;
+		}
+
+		.Combobox__SecondaryText {
+			margin-right: 0;
+		}
+
+		&:focus,
+		&.Combobox_Option__Focus {
+			outline: none;
+			@include __theme-color($white, $gray-dark);
+			@include __theme-background($green-dark, white);
+
+			span {
+				@include __theme-color($white, $gray-dark);
+			}
+
+			svg path {
+				@include __theme-fill($white, $gray-dark);
+			}
+		}
+
+		&.Combobox_Option__Striped:nth-child(even):not(:hover):not(:focus):not(.Combobox_Option__Focus) {
+			@include theme-background-gray2-else-gray2row;
+		}
+	}
+}
