--- src/components/inputs/FlySearchSelect/FlySearchSelectV2.stories.mdx
@@ -1,238 +1,0 @@
-import FlySelect from "./FlySearchSelectV2";
-import Avatar from "../../media/Avatar/Avatar";
-import { Meta, Story, Canvas } from "@storybook/addon-docs/blocks";
-import { Tooltip } from "../../overlays/Tooltip/Tooltip";
-
-<Meta title="Inputs/FlySearchSelectV2" component={FlySelect} />
-
-## Basic (Options = Array)
-
-<Canvas>
-	<Story name="Options = Array">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				inputHeight='s'
-				options={["Test 1", "Test 2"]}
-				onChange={() => console.log("onChange")}
-				value="Test 1"
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Basic (Disabled)
-
-<Canvas>
-	<Story name="Disabled">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				options={["Test 1", "Test 2"]}
-				onChange={() => console.log("onChange")}
-				disabled
-				placeholder="Select an option..."
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Basic (Disabled with Tooltip)
-
-<Canvas>
-	<Story name="Disabled with tooltip">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				container={{
-					element: (
-						<Tooltip
-							content={"This combobox is disabled"}
-							popperOffsetModifier={{ offset: [-55, 10] }}
-							showDelay={300}
-						/>
-					),
-				}}
-				options={["Test 1", "Test 2"]}
-				onChange={() => console.log("onChange")}
-				disabled
-				placeholder="Select an option..."
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Basic (Invalid)
-
-<Canvas>
-	<Story name="Invalid">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				invalid
-				options={["Test 1", "Test 2"]}
-				onChange={() => console.log("onChange")}
-				placeholder="Select an option..."
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Basic (Invalid with message)
-
-<Canvas>
-	<Story name="Invalid with message">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				invalid
-				options={["Test 1", "Test 2"]}
-				onChange={() => console.log("onChange")}
-				placeholder="Select an option..."
-				invalidMessage="Sample error message"
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Basic (Options = Object)
-
-<Canvas>
-	<Story name="Options = Object">
-		<div style={{ width: "90%", height: "125px" }}>
-			<FlySelect
-				options={{
-					a: "Example A",
-					b: "Example B",
-				}}
-				onChange={() => console.log("onChange")}
-				value="b"
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Option groups with striped rows:
-
-<Canvas>
-	<Story name="Striped">
-		<div style={{ width: "90%", height: "310px" }}>
-			<FlySelect
-				optionGroups={{
-					active: {
-						label: "Active Items",
-					},
-					inactive: {
-						label: "Inactive Items",
-					},
-				}}
-				options={{
-					a: {
-						label: "Item A",
-						optionGroup: "active",
-					},
-					b: {
-						disabled: true,
-						label: "Item B",
-						optionGroup: "inactive",
-						secondaryText: "(deactivated)",
-					},
-					c: {
-						label: "Item C",
-						optionGroup: "active",
-					},
-					d: {
-						label: "Item D",
-						optionGroup: "active",
-					},
-					e: {
-						label: "Item E",
-						optionGroup: "active",
-					},
-				}}
-				onChange={() => console.log("onChange")}
-				placeholder="Select Something!"
-				striped={true}
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-## Complex Example:
-
-<Canvas>
-	<Story name="Connect">
-		<div style={{ width: "90%", height: "600px" }}>
-			<FlySelect
-				optionHeight='l'
-				optionGroups={{
-					active: {
-						label: "Active Items",
-					},
-					inactive: {
-						label: "Inactive Items",
-					},
-					connected: {
-						label: "Connected sites",
-					},
-				}}
-				options={{
-					vi6tamon66: {
-						label: 'EduardoTest',
-						optionGroup: 'connected',
-						disabled: true,
-						secondaryText: 'EduardoTest',
-					},
-					f9tl99u20j: {
-						label: 'gameovertest',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v2.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					xvvkshuge5: {
-						label: 'Hot Thoughts',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v3.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					eb4048qhg3: {
-						label: "Let's Learn Local",
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v2.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					xljat07he5: {
-						label: 'Local by WPE',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v1.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					'9tk3oojmil': {
-						label: 'Product Roadmaps',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v1.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					'47b1364yva': {
-						label: 'Spacemind 5000',
-						icon: 'https:flywheel-app-production.s3.amazonaws.com/uploads/site/657948/1643221998-57-0001-1961/expanded_brain_large.jpeg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					'8eht8gi4t8': {
-						label: 'Star Map',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v2.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-					vwddlo8lal: {
-						label: 'Starship Wingnut',
-						icon: 'https://app.getflywheel.com/assets/images/app/avatars/site-v2.svg',
-						optionGroup: undefined,
-						disabled: false,
-					},
-				}}
-				onChange={() => console.log("onChange")}
-				placeholder="Select Something!"
-				striped={true}
-			/>
-		</div>
-	</Story>
-</Canvas>
-
-
